added makePageNumber and re-organizing scripts
